convert gripper subtask to planning_pipeline
This relies on `MoveGroupInterface::constructMotionPlanRequest()` which
is not yet upstream.

This change is required to directly pass the relevant planning scene
in there instead of sending *just the start RobotState* to the action.

This allows subtasks to change other aspects of the scene, e.g.
allowed collisions or new collision objects.