fixup! rename "goal" to "direction"
"Goal" implies a motion to a target configuration.
MoveRelative explicitly does not do that.

"Direction" is usually not used for rotations,
but perfectly valid to describe them.

I merged 544f57416694fb45cc5bb02bde4c3ac34d57bbc3
together with the first version of this rename
that got force-pushed because it was incomplete.