rework RViz panel
There's no need for individually resizable columns, but all numeric
content should autoresize and be visible all the time.

We will add an option to hide the time column in the future.

Remove intermediate AutoAdjustingTreeView layer to simplify code.
Resizing the columns did not work perfectly well before
and the auto_hide_cols_ feature was not really used.