rework PipelinePlanner creation
https://github.com/ros-planning/moveit/pull/2127 merged support
for multiple planner plugins. As a result, the planner specification
should include a pipeline name from now on.

Support the new schema and provide a fallback for older configurations.

While doing so, I moved the create methods to the class they actually belong to
and introduced a clean options parameter