accept const-refs in liftModifiedSolution
clang-tidy rightfully complained that moving rvalues doesn't make
sense when the underlying method accepts const refs.

The rvalue parameter interface, as it is used with spawn/etc. indicates
that the solution is submitted to the system. But with the shared_ptr interface
we need for `liftModifiedSolution` it does not make any sense, especially because
we can't even move the passed pointer to the internal datastructures but have to copy...