switch CI ccov job to Debug build
To provide accurate coverage information.

It was originally RelWithDebInfo because MoveIt uses it
and Travis's timeout would kick in with the old setup.

According to Robert the timeout should not be a problem anymore with
the GHA ci setup. Actually I don't think the timeout was ever a problem for MTC,
but the config was just copied.