Do not replaceImpl when it already serves the right interface
There should be no harm in running init multiple times
and it actually happens with `Task.init(); ...; Task.plan();`.

However, the logic here overwrites impl with a new type by moving
the old pimpl object a couple of times with operator= through inheritance.
This is eventually caught by StagePrivate::operator=() asserting that the typeid is not identical.